pyproject.toml
[project]
name = "fastapi-azure-ai-governance-starter"
version = "0.1.0"
description = "Production-grade FastAPI service on Azure with Key Vault, Managed Identity, OTel, and Terraform IaC"
requires-python = ">=3.12"
readme = "README.md"
license = { text = "MIT" }

dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "sqlalchemy[asyncio]>=2.0.30",
    "asyncpg>=0.29.0",
    "alembic>=1.13.1",
    "python-jose[cryptography]>=3.3.0",
    "httpx>=0.27.0",
    "opentelemetry-sdk>=1.24.0",
    "opentelemetry-instrumentation-fastapi>=0.45b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.45b0",
    "opentelemetry-instrumentation-httpx>=0.45b0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.24.0",
    "azure-monitor-opentelemetry-exporter>=1.0.0b24",
    "azure-keyvault-secrets>=4.8.0",
    "azure-identity>=1.16.0",
    "structlog>=24.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "anyio>=4.4.0",
    "ruff>=0.4.7",
    "mypy>=1.10.0",
    "types-python-jose>=3.3.4",
    "pre-commit>=3.7.0",
    "python-dotenv>=1.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "ANN", "S", "B", "A", "C4", "PT", "RUF"]
ignore = ["ANN101", "ANN102", "ANN401", "S101"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101", "ANN"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=app --cov-report=term-missing --cov-fail-under=70"

[tool.coverage.run]
omit = ["tests/*", "migrations/*"]
